better project folders structure
